logout item added, old menu cleanup
Now uninstalled aps don't show up till the system is rebooted. Also gnome or mate can be logged out of.
If people start using other desktops the graphical environment test code is ready to be added from fluxable-newmenu, but it is not added to keep things
as light as possible.